ugc_app/requirements.txt
aioredis==1.3.1
elasticsearch[async]==7.9.1
fastapi==0.91.0
orjson==3.8.6
pydantic==1.10.2
uvicorn==0.19.0
uvloop==0.17.0
gunicorn==20.1.0
httptools==0.5.0
aiokafka==0.8.0
python-jose==3.3.0
sentry-sdk==1.16.0
python-dotenv==0.21.0
motor==3.1.1
python-logstash-async==2.5.0